Minor convenience stuff

+++ Microsoft.PowerShell_profile.ps1
@@ -1,7 +1,14 @@
 # Make sure we don't conflict with the proper wget command.
-Remove-Item Alias:wget
+If (Test-Path Alias:wget) {
+	Remove-Item Alias:wget
+}
 
 # Handy little alias to use as an alternative to bash's &
 Function bg() {
     Start-Process -NoNewWindow @args
+}
+
+# Since we always want to start Emacs in the background...
+Function emacs() {
+	bg emacs @args
 }
